+++ NEWS.md
@@ -1,7 +1,9 @@
 ### crunch 1.18.1 (under development)
 * `crunchbox` now allows you to specify colors for branding or even category-specific coloring.
 * `makeWeight` allows you to generate new weighting variables based on categorical variables (#80).
+* `cut`, equivalent to `base::cut`, allows you to generate a derived categorical variable based on a numeric variable (#93). 
 * Create a new Crunch dataset from a file by calling `newDataset` directly instead of `newDatasetFromFile`. Also, you can now create a dataset from a hosted file passing its URL to `newDataset(FromFile)`.
+* `as.data.frame()` method for `VariableCatalog` for a view of variable metadata (#75)
 * RStudio users will now be prompted for their password on `login` in a way that conceals the input.
 * Changed the behavior of `changeCategoryID()` to only update numeric values of the category having its id changed when the id and the numeric value are the same.
 * The `autorollback` argument of `appendDataset()` has been deprecated. The option no longer has any effect and a warning will be printed to notify users about the deprecation.